pyproject.toml
[project]
name = "youtube-english-drf"
version = "0.1.0"
description = "Add your description here"
dependencies = [
    "django>=5.0.7",
    "djangorestframework>=3.15.2",
    "django-cors-headers>=4.4.0",
    "psycopg2-binary>=2.9.9",
]
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/youtube_english_drf"]
